remove noop ui, replace in adapter and ansible tests with sdk's TestUi
