+++ packer-plugin-sdk/packer/ui.go
@@ -151,6 +151,56 @@ func (rw *BasicUi) TrackProgress(src string, currentSize, totalSize int64, strea
 	return rw.PB.TrackProgress(src, currentSize, totalSize, stream)
 }
 
+// Safe is a UI that wraps another UI implementation and
+// provides concurrency-safe access
+type SafeUi struct {
+	Sem chan int
+	Ui  Ui
+	PB  getter.ProgressTracker
+}
+
+var _ Ui = new(SafeUi)
+
+func (u *SafeUi) Ask(s string) (string, error) {
+	u.Sem <- 1
+	ret, err := u.Ui.Ask(s)
+	<-u.Sem
+
+	return ret, err
+}
+
+func (u *SafeUi) Say(s string) {
+	u.Sem <- 1
+	u.Ui.Say(s)
+	<-u.Sem
+}
+
+func (u *SafeUi) Message(s string) {
+	u.Sem <- 1
+	u.Ui.Message(s)
+	<-u.Sem
+}
+
+func (u *SafeUi) Error(s string) {
+	u.Sem <- 1
+	u.Ui.Error(s)
+	<-u.Sem
+}
+
+func (u *SafeUi) Machine(t string, args ...string) {
+	u.Sem <- 1
+	u.Ui.Machine(t, args...)
+	<-u.Sem
+}
+
+func (u *SafeUi) TrackProgress(src string, currentSize, totalSize int64, stream io.ReadCloser) (body io.ReadCloser) {
+	u.Sem <- 1
+	ret := u.Ui.TrackProgress(src, currentSize, totalSize, stream)
+	<-u.Sem
+
+	return ret
+}
+
 // NoopProgressTracker is a progress tracker
 // that displays nothing.
 type NoopProgressTracker struct{}
