+++ provisioner/ansible/provisioner.go
@@ -30,7 +30,6 @@ import (
 	"golang.org/x/crypto/ssh"
 
 	"github.com/hashicorp/hcl/v2/hcldec"
-	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/adapter"
 	"github.com/hashicorp/packer/packer-plugin-sdk/common"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep/commonsteps"
@@ -433,7 +432,7 @@ func (p *Provisioner) setupAdapter(ui packersdk.Ui, comm packersdk.Communicator)
 		return "", err
 	}
 
-	ui = &packer.SafeUi{
+	ui = &packersdk.SafeUi{
 		Sem: make(chan int, 1),
 		Ui:  ui,
 	}
